fix: unify csv code and fix duplicate csv export button/popover on sql runner page
