+++ packages/backend/src/routers/projectRouter.ts
@@ -28,7 +28,9 @@ import {
     isAuthenticated,
     unauthorisedInDemo,
 } from '../controllers/authentication';
+import { CsvService } from '../services/CsvService/CsvService';
 import {
+    csvService,
     dashboardService,
     projectService,
     s3Service,
@@ -259,55 +261,13 @@ projectRouter.post(
                 metricQuery.additionalMetrics,
                 metricQuery.tableCalculations,
             );
-            // Ignore fields from results that are not selected in metrics or dimensions
-            const selectedFieldIds = [
-                ...body.metrics,
-                ...body.dimensions,
-                ...body.tableCalculations.map((tc: any) => tc.name),
-            ];
-            const csvHeader = Object.keys(results.rows[0])
-                .filter((id) => selectedFieldIds.includes(id))
-                .map((id) => getItemLabel(itemMap[id]));
-            const csvBody = results.rows.map((row) =>
-                Object.keys(row)
-                    .filter((id) => selectedFieldIds.includes(id))
-                    .map((id) => {
-                        const rowData = row[id];
-                        const item = itemMap[id];
-                        if (
-                            isField(item) &&
-                            item.type === DimensionType.TIMESTAMP
-                        ) {
-                            return moment(rowData.value.raw).format(
-                                'YYYY-MM-DD HH:mm:ss',
-                            );
-                        }
-                        if (isField(item) && item.type === DimensionType.DATE) {
-                            return moment(rowData.value.raw).format(
-                                'YYYY-MM-DD',
-                            );
-                        }
-                        if (onlyRaw) {
-                            return rowData.value.raw;
-                        }
-                        return rowData.value.formatted;
-                    }),
-            );
 
-            const csvContent: string = await new Promise((resolve, reject) => {
-                stringify(
-                    [csvHeader, ...csvBody],
-                    {
-                        delimiter: ',',
-                    },
-                    (err, output) => {
-                        if (err) {
-                            reject(new Error(err.message));
-                        }
-                        resolve(output);
-                    },
-                );
-            });
+            const csvContent: string = await CsvService.convertApiResultsToCsv(
+                results,
+                onlyRaw,
+                metricQuery,
+                itemMap,
+            );
 
             const fileId = `csv-${nanoid()}.csv`;
 
@@ -713,37 +673,9 @@ projectRouter.post(
                     req.body.sql,
                 );
 
-            const csvHeader = Object.keys(results.rows[0]);
-            const csvBody = results?.rows.map((row) =>
-                Object.values(results?.fields).map((field, fieldIndex) => {
-                    if (field.type === DimensionType.TIMESTAMP) {
-                        return moment(Object.values(row)[fieldIndex]).format(
-                            'YYYY-MM-DD HH:mm:ss',
-                        );
-                    }
-                    if (field.type === DimensionType.DATE) {
-                        return moment(Object.values(row)[fieldIndex]).format(
-                            'YYYY-MM-DD',
-                        );
-                    }
-                    return Object.values(row)[fieldIndex];
-                }),
+            const csvContent = await CsvService.convertSqlQueryResultsToCsv(
+                results,
             );
-
-            const csvContent: string = await new Promise((resolve, reject) => {
-                stringify(
-                    [csvHeader, ...csvBody],
-                    {
-                        delimiter: ',',
-                    },
-                    (err, output) => {
-                        if (err) {
-                            reject(new Error(err.message));
-                        }
-                        resolve(output);
-                    },
-                );
-            });
 
             const fileId = `csv-${nanoid()}.csv`;
